lora: Fix issue with CRC Error IRQ bit staying set
Modem won't ever clear this in continuous RX mode, so needs to be cleared by the
driver. Had previously held off on clearing "non-final" IRQ bits, in the hope
that would avoid any race for an error present in the next packet.

Hopeful fix for issue reported by Barry on Discord.

Signed-off-by: Angus Gratton <angus@redyak.com.au>